+++ CHANGELOG.md
@@ -62,3 +62,4 @@ Upcoming changes compared to [NetCoreDbg](https://github.com/Samsung/netcoredbg)
 - Fixed bug in corhost related logic (TPA list creation).
 - Fixed some methods `void *&` (`PVOID &`) parameters to `void **`.
 - Fixed stacktrace for exception in async methods (exception rethrow with `System.Runtime.ExceptionServices.ExceptionDispatchInfo.Throw()`).
+- Fixed Cancel Request, `requestId` is optional parameter now.